java chinese iter_7

+++ src/main/java/org/636766fe1a6d9265ec017833/636766fe1a6d9265ec017833.java
@@ -14,16 +14,36 @@ public class FileUtils {
             throw new NullPointerException("File cannot be null");
         }
 
-        file.deleteOnExit();
+        if (!file.exists()) {
+            return;
+        }
 
         if (file.isDirectory()) {
-            File[] files = file.listFiles();
-            if (files != null) {
-                for (File subFile : files) {
-                    forceDeleteOnExit(subFile);
-                }
+            deleteDirectoryOnExit(file);
+        } else {
+            file.deleteOnExit();
+        }
+    }
+
+    private static void deleteDirectoryOnExit(File directory) throws IOException {
+        if (!directory.isDirectory()) {
+            throw new IllegalArgumentException("Not a directory: " + directory);
+        }
+
+        File[] files = directory.listFiles();
+        if (files == null) {
+            throw new IOException("Failed to list contents of " + directory);
+        }
+
+        for (File file : files) {
+            if (file.isDirectory()) {
+                deleteDirectoryOnExit(file);
+            } else {
+                file.deleteOnExit();
             }
         }
+
+        directory.deleteOnExit();
     }
 
     public static void main(String[] args) {
